// apps/admin/src/features/AddNewAddressModal/types/addressesTypes.ts
export interface AddressType {
  id: number
  city: string
  street: string
  house: string
  entrance: string
  floor: string
  apartment: string
  intercom: string
}

export interface AddressFormSchema {
  city: string
  street: string
  house: string
  entrance: string
  floor: string
  apartment: string
  intercom: string
}